Wasn't able to fix the timing issue. Revert single-root-UI and add a preset version of it
