+++ test/extension-tests/single-root-UI-presets/project-folder/CMakePresets.json
@@ -1,0 +1,18 @@
+{
+  "version": 2,
+  "configurePresets": [
+    {
+      "name": "Linux1",
+      "description": "Sets generator, build and install directory, vcpkg",
+      "generator": "Ninja",
+      "binaryDir": "${workspaceFolder}/build",
+      "cacheVariables": {
+        "CMAKE_BUILD_TYPE": "Debug",
+        "CMAKE_TOOLCHAIN_FILE": {
+            "type": "FILEPATH",
+            "value": "${workspaceFolder}/test-toolchain.cmake"
+        }
+      }
+    }
+  ]
+}
