+++ test/extension-tests/single-root-UI-presets/project-folder/TestCacheInit.cmake
@@ -1,0 +1,1 @@
+set(CMT_COOKIE cache-init-cookie CACHE STRING "Cache-init forced cookie value")
