+++ test/extension-tests/single-root-UI-presets/project-folder/test-toolchain.cmake
@@ -1,0 +1,2 @@
+# This file is only here as a test toolchain!
+message("Hello from test-toolchain.cmake")
