+++ test/extension-tests/single-root-UI-presets/test/debugger.test.ts
@@ -1,13 +1,11 @@
-import {DefaultEnvironment, expect, getFirstSystemKit} from '@test/util';
+import {DefaultEnvironment, expect} from '@test/util';
 // import sinon = require('sinon');
 import * as vscode from 'vscode';
-import CMakeTools from '@cmt/cmake-tools';
 
 // tslint:disable:no-unused-expression
 
-suite('[Debug/Launch interface]', async () => {
+suite('[Debug/Launch interface using Presets]', async () => {
   let testEnv: DefaultEnvironment;
-  let cmakeTools: CMakeTools;
 
   setup(async function(this: Mocha.Context) {
     this.timeout(100000);
@@ -15,13 +13,13 @@ suite('[Debug/Launch interface]', async () => {
     const build_loc = 'build';
     const exe_res = 'output.txt';
 
-    testEnv = new DefaultEnvironment('test/extension-tests/single-root-UI/project-folder', build_loc, exe_res);
-    cmakeTools = await CMakeTools.create(testEnv.vsContext, testEnv.wsContext);
-
-    const kit = await getFirstSystemKit(cmakeTools);
-    console.log("Using following kit in next test: ", kit);
-    await vscode.commands.executeCommand('cmake.setKitByName', kit.name);
+    testEnv = new DefaultEnvironment('test/extension-tests/single-root-UI-presets/project-folder', build_loc, exe_res);
     testEnv.projectFolder.buildDirectory.clear();
+
+    await vscode.commands.executeCommand('cmake.setConfigurePreset', 'LinuxUser1');
+    await vscode.commands.executeCommand('cmake.setBuildPreset', '__defaultBuildPreset__');
+    await vscode.commands.executeCommand('cmake.setTestPreset', '__defaultTestPreset__');
+
     expect(await vscode.commands.executeCommand('cmake.build')).to.be.eq(0);
   });
 
@@ -39,4 +37,3 @@ suite('[Debug/Launch interface]', async () => {
     // sinon.assert.calledWith(testEnv.vs_debug_start_debugging);
   }).timeout(60000);
 });
-
